Backend/requirements.txt
Flask
Flask-Cors
python-dotenv
pdfplumber
python-docx
spacy
nltk
openai
google-generativeai